move instrumentation to separate module
